Change in SpritesheetCache
* SpritesheetCache now returns only IDs on load and create empty
* Spritesheet reference is retrieved only through getSpritesheet(handle) method
* Sweeping changes to conform to this new interface
* Change the delta time argument type in Actions from double to float
* Change the State type that is being saved from Spritesheet& to uint ID
* Remove logging in InputEngine in case of unprocessed event type

+++ source/Core/Action/ActionLogic.h
@@ -15,4 +15,4 @@ namespace Core
 	class GameContext;
 }
 
-typedef void(*ActionLogic)(double dt, Core::GameContext& context, std::set<Core::Action*>& actions);
+typedef void(*ActionLogic)(float dt, Core::GameContext& context, std::set<Core::Action*>& actions);
